requirements.txt
panel~=1.3.8
param~=2.0.2
webdriver-manager~=4.0.1
pytest~=8.0.0
praw ~=7.7.1
python-dotenv~=1.0.1